lastName validation changed, passport middlewares moved to middleware folder
