+++ server/config/passport.js
@@ -1,5 +1,5 @@
-const localStrategy = require('../strategies/local');
-const jwtStrategy = require('../strategies/jwt');
+const localStrategy = require('../middleware/strategies/local');
+const jwtStrategy = require('../middleware/strategies/jwt');
 //If we want to add  strategies or functions like serialize and deserialize we use this function
 const passportConfig = passport => {
 	passport.use('local', localStrategy);
